Fix exception messages and XML Driver
